Language and theme choices
Instead of browserifying all of highlight.js (176 languages), we're switching to a vendored build that only has 8 languages by default, and adding the option of letting the user provide their own build if they want a different set.

While I was at it, it was easy to also add an option to override the included theme, so we have that too.